requirements.txt
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
pigpio==1.78
pyserial==3.5
